+++ src/UI/GUIPopupMenuGTK.cpp
@@ -15,7 +15,7 @@ void UI::GUIPopupMenu::ShowMenu(UI::GUIControl *ctrl, OSInt x, OSInt y)
 	{
 		gint tranX;
 		gint tranY;
-		if (gtk_widget_translate_coordinates((GtkWidget*)ctrl->GetHandle(), (GtkWidget*)frm->GetHandle(), x, y, &tranX, &tranY))
+		if (gtk_widget_translate_coordinates((GtkWidget*)ctrl->GetHandle(), (GtkWidget*)frm->GetHandle(), (gint)x, (gint)y, &tranX, &tranY))
 		{
 			x = tranX;
 			y = tranY;
@@ -26,8 +26,8 @@ void UI::GUIPopupMenu::ShowMenu(UI::GUIControl *ctrl, OSInt x, OSInt y)
 	GdkWindow *window = gtk_widget_get_window((GtkWidget*)ctrl->GetHandle());
 
 	GdkRectangle rect;
-	rect.x = x;
-	rect.y = y;
+	rect.x = (int)x;
+	rect.y = (int)y;
 	rect.width = 0;
 	rect.height = 0;
 	gtk_menu_popup_at_rect((GtkMenu*)this->hMenu, window, &rect, GDK_GRAVITY_SOUTH_WEST, GDK_GRAVITY_NORTH_WEST, 0);
